requirements.txt
opencv-python-headless
gunicorn
gevent
flask
requests
numpy
pillow
google-genai
python-dotenv
